Point Of Sales System Application
Point of sale (POS), a critical piece of a point of purchase, refers to the place where a customer executes the payment for goods or services and where sales taxes may become payable. It can be in a physical store, where POS terminals and systems are used to process card payments or a virtual sales point such as a computer or mobile electronic device.